Improving execute()
